+++ playground/entries/Chart/Line.tsx
@@ -1,0 +1,86 @@
+import { Chart } from '@semcore/d3-chart';
+import { curveBumpX, curveCardinal, curveLinearClosed } from 'd3-shape';
+import React from 'react';
+
+import ChartControls from './common/controls';
+import type { CommonChartProps, LegendChartProps } from './common/controls';
+import type { JSXProps } from '../../types/JSXProps';
+import type { PlaygroundEntry } from '../../types/Playground';
+import createGithubLink from '../../utils/createGHLink';
+
+const Curves = ['curveCardinal', 'curveLinearClosed', 'curveBumpX'] as const;
+
+type LineChartProps = {
+  commonChartProps: CommonChartProps;
+  legendProps: LegendChartProps;
+  linearChartProps: {
+    curve: (typeof Curves)[number];
+    showDots: boolean;
+  };
+};
+
+const CurveMap = {
+  curveCardinal,
+  curveLinearClosed,
+  curveBumpX,
+};
+
+export type LineChartJSXProps = JSXProps<LineChartProps>;
+
+const data = [...Array(5).keys()].map((d, i) => ({
+  x: i,
+  Line1: Math.random() * 10,
+  Line2: Math.random() * 10,
+  Line3: Math.random() * 10,
+}));
+
+function getJSX(props: LineChartJSXProps) {
+  const { curve, ...restLinearChartProps } = props.linearChartProps;
+
+  return (
+    <Chart.Line
+      plotWidth={300}
+      plotHeight={200}
+      groupKey='x'
+      data={data}
+      aria-label='Line chart'
+      {...props.commonChartProps}
+      {...restLinearChartProps}
+      {...(props.legendProps && {
+        legendProps: props.legendProps,
+        showLegend: props.commonChartProps.showLegend as true,
+      })}
+      {...(props.legendProps?.patterns && { patterns: props.legendProps.patterns })}
+      {...(CurveMap[curve] && { curve: CurveMap[curve] })}
+    />
+  );
+}
+
+const entry: PlaygroundEntry<LineChartJSXProps> = {
+  JSX: (props) => getJSX(props),
+  controls: {
+    linearChartProps: {
+      type: 'group',
+      groupName: 'Linear chart props',
+      isOpenedByDefault: true,
+      controls: {
+        curve: {
+          type: 'select',
+          value: 'No curve',
+          options: ['No curve', ...Curves],
+          displayName: 'Curve',
+        },
+        showDots: {
+          type: 'boolean',
+          value: true,
+          displayName: 'Dots',
+        },
+      },
+    },
+    ...ChartControls,
+  },
+  link: createGithubLink('d3-chart'),
+  filterProps: ['data'],
+};
+
+export default entry;