+++ playground/events/events.ts
@@ -1,0 +1,4 @@
+export default {
+  CopyCodeBtnClick: 'playground:copy-code-btn-click',
+  ViewSourceBtnClick: 'playground:view-source-btn-click',
+};