+++ semcore/core/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.1.1] - 2025-08-29
+
+### Changed
+
+- FocusLock work when popper opened without focus and trigger has few focusable elements.
+
+## [16.1.0] - 2025-07-24
+
+### Added
+
+- New tokens for FeaturePopover's new theme styles: `feature-popover-bg`, `feature-popover-bg-neutral`, `feature-popover-dot-neutral`, `feature-popover-dot-neutral-outer-border`.
+
 ## [16.0.4] - 2025-07-07
 
 ### Added