+++ semcore/data-table/CHANGELOG.md
@@ -8,6 +8,19 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - `Data` type to `cellRender` function.
 
+### Fixed
+
+- Unnecessary re-render when data was updated but old values were still used.
+- Shadows remain after resizing table width.
+- `onSortChange` is called twice when click is on a sort icon.
+- Hover behaviour for sort icon in case the content is justified to the right.
+
+### Changed
+
+- Type description for `DataTableProps`/`CellRenderProps`.
+- `AccordionToggle`'s `margin-left`.
+- `DataTable` display to `block` & `DataTable.Header` display to `grid` when `data` is empty.
+
 ## [16.1.0] - 2025-08-08
 
 ### Added