+++ semcore/pagination/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.1.7] - 2025-08-29
+
+### Changed
+
+- Type description for `PaginationProps`.
+
 ## [16.1.6] - 2025-08-07
 
 ### Fixed